UPSTREAM: AMD CIMx SB800: late.c: Use variable `device` from for loop condition
Use the variable `device` instead of `dev` in the predicate of
the if condition, as `dev` is not changed in the for loop.

The for loop was added in the following commit.

    commit 8fed77ae4c46122859d0718678e54546e126d4bc
    Author: Scott Duplichan <scott@notabs.org>
    Date:   Sat Jun 18 10:46:45 2011 -0500

        ASRock E350M1: Configure SB800 GPP ports to support onboard pcie nic

        Reviewed-on: http://review.coreboot.org/44

The assumption that the devices are ordered in the tree seem to
hold in this case (although it is not ensured) and therefore at
least with the ASRock E350M1 no (visible) change is experienced as
the children are all of type `DEVICE_PATH_PCI`.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ic6fdaad7eab53d769d6b288f296d6ecfd1bcb7da
Signed-off-by: Martin Roth <martinroth@google.com>
Original-Commit-Id: e4a016ff17b455a8acd0ea72b30caabb622ce224
Original-Change-Id: Iaa2fa13305dbe924965d27680cd02fe30c2f58a5
Original-Signed-off-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Signed-off-by: Kysti Mlkki <kyosti.malkki@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/2562
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/868757
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Martin Roth <martinroth@chromium.org>