UPSTREAM: payloads/external/tianocore: build UEFI payload
Update the existing tianocore payload (which didn't do any more than
adding an elf payload with a specific name) to fetch and build the
UEFI corebootPayloadPackage, using the coreboot toolchain for
compilation. Only checkout the commit when changing commit IDs or if
version is master, instead of every time it builds.

Currently working if patches are merged into the upstream edk2
repository (to be included in a follow-on patch).

BUG=none
BRANCH=none
TEST=none

Change-Id: I1b015de5bc7639a90f557b957cad85c222449535
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 4769cc3c0cc0ef465484013e8b320df79b0cafac
Original-Change-Id: I0bf4cedec2d6821ae2a04184ebb5cf88979ccee3
Original-Signed-off-by: Martin Roth <martinroth@google.com>
Original-Signed-off-by: Philipp Deppenwiese <zaolin@das-labor.org>
Original-Signed-off-by: Evelyn Huang <evhuang@google.com>
Original-Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/15057
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/610604
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>