UPSTREAM: cpu/x86/smm: Fix explicit 'addr32' usage in clang builds
The addr32 prefix is required by binutils, because even when
given an explicit address which is greater than 64KiB, it will
throw a warning about truncation, and stupidly emit the opcode
with a 16-bit addressing mode and the wrong address.

However, in the case of LLVM, this doesn't happen, and is happy
to just use 32-bit addressing whenever it may require it. This
means that LLVM never really needs an explicit addr32 prefix to
use 32-bit addressing in 16-bit mode.

BUG=none
BRANCH=none
TEST=none
CQ-DEPEND=CL:650708

Change-Id: Ia2824b25175d990863ca6d16e6aa4351b6be1be0
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 1104c278e396860567c4d20ff2b0868fa6b02696
Original-Change-Id: Ia160d3f7da6653ea24c8229dc26f265e5f15aabb
Original-Also-by: Damien Zammit <damien@zamaudio.com>
Original-Signed-off-by: Damien Zammit <damien@zamaudio.com>
Original-Reviewed-on: https://review.coreboot.org/21219
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-by: Philippe Mathieu-Daud <f4bug@amsat.org>
Reviewed-on: https://chromium-review.googlesource.com/648941
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>