UPSTREAM: util/bucts: Add tool to manipulate BUC.TS bit on Intel targets
The purpose of this tool is to manipulate and get information about the
`Back Up Control, Top Swap` mechanism present on most Intel Southbridges.

This tool is initially written by Peter Stuge.

This tool makes it possible to have a backup mechanism for the
bootblock by using the southbridges Back Up Control Top Swap.

Sometimes it is also possible to circumvent vendor write protection mechanisms
in order to flash coreboot. An example of where this would be useful would be
the Lenovo Thinkpad X60 and T60.

BUG=none
BRANCH=none
TEST=none

Change-Id: I1e78f3d2d45ac2eae24ce3580630aa338b142828
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: a29498f59952e89e858affde52c3f7565ab7f210
Original-Change-Id: I12cc2e91396f096fc979e23848e1929cb6c44fc5
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/18224
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Ignore-CL-Reviewed-on: https://review.coreboot.org/28775
Ignore-CL-Reviewed-on: https://review.coreboot.org/29030
Ignore-CL-Reviewed-on: https://review.coreboot.org/29269
Reviewed-on: https://chromium-review.googlesource.com/1341427
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>