UPSTREAM: AGESA: Move romstage-ramstage splitline
In AGESA specification AmdInitEnv() is to be called once
host memory allocator has started. In coreboot context this
could mean either availability of CBMEM or malloc heap.

As for AmdS3LateRestore(), there is no requirement to have
it run as part of the romstage either.

BUG=none
BRANCH=none
TEST=none

Change-Id: I2bb49726ea6d9417fefbf46ff9a83ec26c4dbc39
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 21e609c1c92c47bd8990c2ce25adf5987b5dd1b2
Original-Change-Id: Icc8d97b82df89e2480e601d5c2e094de0365b0a5
Original-Signed-off-by: Kysti Mlkki <kyosti.malkki@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/18888
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/597587
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>