UPSTREAM: mainboard/hp: Add HP Elitebook 8470p
The code is based on the code generated by autoport.

The EHCI debug port is between the DP port and eSATA port.
The serial port on dock can also be used for debugging.

The model with discrete graphics is not tested.

Tested and work:
- memory: 0+4G, 4G+0, 4G+4G
- Linux (Arch Linux with kernel 4.10.6) boot from SeaBIOS payload
  with native graphics init
- WLAN
- keyboard, trackpoint and touchpad
- USB
- serial port on dock
- fan control
- AC and battery status

(EC) blobs:

This laptop uses SMSC KBC1126 EC, and there are two blobs needed by it.

You can use the tools in util/kbc1126/ to extract them and insert them
to the coreboot image using the following configuration:

  -> Chipset
    -> Add firmware images for KBC1126 firmware

BUG=none
BRANCH=none
TEST=none

Change-Id: I6ee6d55188e2c851d9ee77a12bdfc465f80832ee
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: df48d8b2c26874196d9d9511408c493ae1f23d13
Original-Change-Id: Icbc051e2272b8ea73627940db15a56901d737472
Original-Signed-off-by: Iru Cai <mytbk920423@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/18985
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-on: https://chromium-review.googlesource.com/624109
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>