UPSTREAM: mb/*/*: Remove rtc nvram configurable baud rate
There have been discussions about removing this since it does not seem
to be used much and only creates troubles for boards without defaults,
not to mention that it was configurable on many boards that do not
even feature uart.

It is still possible to configure the baudrate through the Kconfig
option.

BUG=none
BRANCH=none
TEST=none

Change-Id: I68f2b9cadd154f4816a63b407da9416fbe3e7539
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: b29078e4015bbc3e8cf00ba64f0799c087546563
Original-Change-Id: I71698d9b188eeac73670b18b757dff5fcea0df41
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/19682
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/681035
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>