UPSTREAM: riscv: Update register address
I triggered a bug, when I try to debug riscv code by spike.
This bug is caused by an instruction exception[csrwi 0x320,7].
This is operate for mcounteren. This address is error. 0x306
is right. scounteren is not need to be set, because S-mode
code controls it.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ic28605ee049fc87a5b08a89906e03cc22d5a752f
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: d277960eb8f7c8ed068f1823049c701e00ea8e46
Original-Change-Id: If0bea4bf52d8ad2fb2598724d6feb59dc1b3084a
Original-Signed-off-by: wxjstz<wxjstz@126.com>
Original-Reviewed-on: https://review.coreboot.org/20043
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Jonathan Neuschfer <j.neuschaefer@gmx.net>
Reviewed-on: https://chromium-review.googlesource.com/686938
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>