UPSTREAM: sis/sis966: Clean up sata.c
Wow, this one is disliked by clang for the empty for() loop, but
looking at the file just makes my eyes bleed a little bit. I remember
the circumstances under which we let this code go in. It was supposed
to save contributions from a vendor, but that never worked out.
Just to keep the little chunks down, here's an indent run and some of
the cruft removed that doesn't actually contribute to functionality in
any way.

BUG=none
BRANCH=none
TEST=none

Change-Id: I247cf19f48ac018969356fdb49d6819e6803905f
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: cb3205d153b7fa80aef4f91c5b5aa3ba636dc777
Original-Change-Id: Ie82166ca82f09c4b66decfde5ad194a2d70b0708
Original-Signed-off-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Original-Reviewed-on: https://review.coreboot.org/20386
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/551667
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>