UPSTREAM: crossgcc: patch out uses of $(P) variable from gcc/ada build
Gentoo likes to use that variable for itself and insists on keeping it.
Meanwhile it doesn't seem to be set or used anywhere else in the gcc
build, and it seems there was a big $(P)-pruning going on in 2000,
so why is it even (still) there?

Related upstream change can be found at
https://gcc.gnu.org/ml/gcc-patches/2017-07/msg01015.html

BUG=none
BRANCH=none
TEST=none

Change-Id: I45866c7635e36e29090ef254f49357c02294d7df
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 9f4864f499864345e5b3296251b0e1d636a04572
Original-Change-Id: I2c2bdf9cb215c489f760f43642a86592924e4e65
Original-Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Reviewed-on: https://review.coreboot.org/20612
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/578361
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>